+++ src/test/resources/fixtures/processing/blade/attributes-class-merge-noop/input.blade.php
@@ -1,0 +1,8 @@
+<div {{ $attributes->class([
+    'text-center p-4 flex bg-red-500 font-bold' => $active,
+    'font-bold flex p-4',
+]) }}></div>
+
+<div {{ $attributes->merge([
+    'class' => 'text-center p-4 flex bg-red-500 font-bold',
+]) }}></div>